entrega realizada

+++ src/components/ProductsList/Product/index.jsx
@@ -1,16 +1,25 @@
 import React, { useEffect } from 'react'
+import { toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 
 import { StyledProducList, StyledDivProducList, StyledDiv2ProducList } from './style'
 import { StyledButton } from '../../../styles/button'
 import { Title_1, Title_2, Text } from '../../../styles/typography'
 
-export function Product({ name, category, price, img, setIsModalOpen, setProductsCart, productsCart }) {
-  function handleOpenModal() {
-    setIsModalOpen(true)
-    const newProduct = { img, name, price }
-    const updatedCart = [...productsCart, newProduct]
-    setProductsCart(updatedCart)
-    localStorage.setItem('productsCart', JSON.stringify(updatedCart))
+export function Product({ name, category, price, img, id, setIsModalOpen, setProductsCart, productsCart }) {
+  function handleOpenModal(){
+    const newProduct = { id, img, name, price }
+    const existinProduct = productsCart.find(product => product.id === id)
+
+    if(existinProduct){
+      toast.error("Esse produto já foi adicionado ao carrinho!");
+      setIsModalOpen(false)
+    }else{
+      setIsModalOpen(true)
+      const updatedCart = [...productsCart, newProduct]
+      setProductsCart(updatedCart)
+      localStorage.setItem('productsCart', JSON.stringify(updatedCart))
+    }
   }
 
   useEffect(() => {
@@ -19,6 +28,7 @@ export function Product({ name, category, price, img, setIsModalOpen, setProduct
       setProductsCart(JSON.parse(savedCart))
     }
   }, [])
+
 
   return (
     <StyledProducList>
